move Trial class to hpo_base.py
